fixed nnrt bug

+++ image_process/depth_anything/depth_imageprocess.cpp
@@ -29,7 +29,7 @@ DepthImageprocess::DepthImageprocess(const ai_framework::Config &config,
 
 void DepthImageprocess::PreProcess(const std::vector<cv::Mat> &input,
                                    void **tenssors) {
-  for (int i = 0; i < input.size(); ++i) {
+  for (size_t i = 0; i < input.size(); ++i) {
     auto &original_input = input.at(i);
     int resize_width = target_size_;
     int resize_height = target_size_;
